Added a basic wire system
The system is essentially a singly linked list. Each node in the list is
responsable for rendering a line to the next node in the list. The idea
is to place these nodes on objects such as the transformer or the
breaker and then link them together in the editor. At runtime the
scripts for the wirenodes will render lines between them. Right now they
only render a basic line, curves and movement(such as swaying in the
wind) will be added later.