Added a simple electrical explosion
I added a particle system to the transformer objects, and a collider
trigger, when the squirrel runs into the collider trigger it will
trigger the particle system and "damage" the transformer. since the
squirrel still can't climb onto components I put the trigger on the
right side of the transformer. Currently the squirrel is unaffected by
running into the transformer