Added a basic switch, started adding functionality via a script
I made a basic switch model and a script to rotate the lever arm. I
started working on a system to pass in 'events' to the lever script that
woudl be triggered when it is interacted with, however it's proving
harder than I expected and isn't functional at the moment. Commiting to
ensure my work is saved.