Added some models for tools, added the trigger for the squirrel event, other small changes
Added a scraper model for use as a tool to remove the squirrel once it
has been fried.
Added a magic wand for general tool debugging(to be used as a stand in
for other tools)
Added an abstract SimulationEvent script which is to be used as a base
for all events in the simulation, it's similar to a interactable however
I felt it best if we seperated the 2 api's. Changed the lever script to
search for a event script attached to itself, also allowed it to be
specified in the editor, however it still needs to be attached to an
object. However events can be coded to be triggered by key presses by
manually specifying them in the inherited event script. Every event will
need it's own script and will need to be inherited from the abstract
class. Also added a basic squirrel event script.