Added a new class to handle configurable input
Added an InputManager class, this class will read in a Input.config file
which has information on id-key pairs, any class wanting input from the
keyboard will request a check by passing an id to this class, it will
return true or false based on if the key assigned to that id is pressed.
It is configurable and can write new Input.config files for changable
input. Can also be dynamically changed while the code is running.