made some changes to the squirrel event
increased the size of the transformer collider,
made some changes to the squirrel follow path so his turning is a little
more reliable.
made it so the squirrel locks his rotation towards the transformer after
collision.
moved the tools so they don't disrupt the squirrel path following.
also made the player taller