added a tail to the squirrel
I added a tail to the squirrel mesh, however this tail doesn't work well
with the fur shader. I think this is a texturing issue, the fur shader
uses the same uv coordinates as the texture and I believe that since I
extruded the tail verticies from the squirrel mesh the uv coordinates
were just stretched over the new face rather than getting their own
coordinates. Basically it's stretching the texture over the tail, so it
doesn't space our the fur correctly.